lots of comments for the core code

+++ backend.py
@@ -1,3 +1,7 @@
+"""
+Provides a protected administrative area for uploading and deleteing images
+"""
+
 import os
 import datetime
 
@@ -11,16 +15,25 @@ from google.appengine.api import users
 from models import Image
 
 class Index(webapp.RequestHandler):
+    """
+    Main view for the application.
+    Protected to logged in users only.
+    """
     def get(self):
-        self.response.out.write('<html><body>')
-
+        "Responds to GET requets with the admin interface"
+        # query the datastore for images owned by
+        # the current user. You can't see anyone elses images
+        # in the admin
         images = Image.all()
         images.filter("user =", users.get_current_user())
         images.order("-date")
 
+        # we are enforcing loggins so we know we have a user
         user = users.get_current_user()
+        # we need the logout url for the frontend
         logout = users.create_logout_url("/")
-        
+
+        # prepare the context for the template
         context = {
             "images": images,
             "logout": logout,
@@ -32,33 +45,57 @@ class Index(webapp.RequestHandler):
         self.response.out.write(template.render(path, context))
 
 class Deleter(webapp.RequestHandler):
+    "Deals with deleting images"
     def post(self):
-       image = db.get(self.request.get("key"))
-       image.delete()
-       self.redirect('/')
+        "Delete a given image"
+        # we get the user as you can only delete your own images
+        user = users.get_current_user()
+        image = db.get(self.request.get("key"))
+        # check that we own this image
+        if image.user == user:
+            image.delete()
+        # whatever happens rediect back to the main admin view
+        self.redirect('/')
        
 class Uploader(webapp.RequestHandler):
+    "Deals with uploading new images to the datastore"
     def post(self):
-        image = Image()
-        original_content = self.request.get("img")
+        "Upload via a multitype POST message"
         
         try:
+            # check we have numerical width and height values
             width = int(self.request.get("width"))
             height = int(self.request.get("height"))
         except ValueError:
+            # if we don't have valid width and height values
+            # then just use the original image
             image_content = self.request.get("img")
         else:
+            # if we have valid width and height values
+            # then resize according to those values
             image_content = images.resize(self.request.get("img"), width, height)
         
+        # get the image data from the form
+        original_content = self.request.get("img")
+        # always generate a thumbnail for use on the admin page
         thumb_content = images.resize(self.request.get("img"), 100, 100)
+        
+        # create the image object
+        image = Image()
+        # and set the properties to the relevant values
         image.image = db.Blob(image_content)
+        # we always store the original here in case of errors
+        # although it's currently not exposed via the frontend
         image.original = db.Blob(original_content)
         image.thumb = db.Blob(thumb_content)
         image.user = users.get_current_user()
                 
+        # store the image in the datasore
         image.put()
+        # and redirect back to the admin page
         self.redirect('/')
                 
+# wire up the views
 application = webapp.WSGIApplication([
     ('/', Index),
     ('/upload', Uploader),
@@ -66,6 +103,7 @@ application = webapp.WSGIApplication([
 ], debug=True)
 
 def main():
+    "Run the application"
     run_wsgi_app(application)
 
 if __name__ == '__main__':
